+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### 🚀 Features
+
+- Add new extra command to install TypeScript declarations for packages having external declarations (e.g. `@types/lodash`).
+
 ## v0.1.1
 
 ### 🚀 Features